requirements.txt
numpy
PyOpenGL
pyrr
imgui
glfw